переписывает store на slices
