+++ src/store/index.ts
@@ -1,11 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer';
+// import { reducer } from './reducer';
 import { createAPI } from '../services/api';
+import { offersSlice } from './slices/offers';
 
 export const api = createAPI();
 
 export const store = configureStore({
-  reducer,
+  reducer: offersSlice.reducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
